Refactor: change modal to singlemovie
