+++ src/components/App/App.js
@@ -3,7 +3,7 @@ import Header from '../Header/Header'
 import MovieContainer from '../MovieContainer/MovieContainer';
 import './App.css';
 import { getAllData } from '../../api-calls';
-import SingleMovie from '../Modal/Modal';
+import SingleMovie from '../SingleMovie/SingleMovie';
 import { Route, Switch } from 'react-router-dom';
 import NotFound from '../NotFound/NotFound';
 import InternalServerError from '../InternalServerError/InternalServerError';
@@ -30,7 +30,7 @@ class App extends Component{
         <Header />
         <Switch>
           <Route exact path='/' render={ () => <MovieContainer movies={this.state.movies}/> } />
-          <Route exact path='/modal/:id' render={ ({match}) => <SingleMovie id={match.params.id} /> } />
+          <Route exact path='/singlemovie/:id' render={ ({match}) => <SingleMovie id={match.params.id} /> } />
           <Route component={NotFound} />
           <Route component={InternalServerError} />
         </Switch>
