+++ src/components/SingleMovie/SingleMovie.css
@@ -27,13 +27,13 @@
     contain: content;
 }
 /* ^^^^^NEED TO REFACTOR TO BE RESPONSIVE WITH PAGE */
-.modal-mini-poster {
+.single-movie-mini-poster {
     position: absolute;
     height: 20vh;
     margin: 15px;
 }
 
-.modal-back-drop{
+.single-movie-back-drop{
     align-items: baseline;
     height: 100vh;
     width: 100vw;
@@ -51,13 +51,13 @@ p {
 }
 
 
-.modal-title{
+.single-movie-title{
     position: absolute;
     height: 5vh;
     margin-left: 12vw;
 }
 
-.modal-overview {
+.single-movie-overview {
     height: 20vh;
     width: 45vw;
     contain: content;
@@ -67,7 +67,7 @@ p {
     flex-wrap: wrap;
 }
 
-.modal-release-date {
+.single-movie-release-date {
     position: absolute;
     margin-top: 8vh;
     margin-left: 12vw;
@@ -98,7 +98,7 @@ p {
     margin-left: 12vw;
 }
 
-.modal-rating {
+.single-movie-rating {
     position: absolute;
     font-size: 2em;
     background-color: white;
